Revert "The function ppl_set_deterministic_timeout() in the C interface now takes an unsigned long long parameter."
This reverts commit 7781b2d72ad31906c830a16a942b31ac305924c2.

+++ interfaces/C/ppl_c_header.h
@@ -420,7 +420,7 @@ ppl_reset_timeout PPL_PROTO((void));
   weight thresholds when upgrading to newer version of the PPL.
 */
 int
-ppl_set_deterministic_timeout PPL_PROTO((unsigned long long weight));
+ppl_set_deterministic_timeout PPL_PROTO((unsigned weight));
 
 /*! \brief
   Resets the deterministic timeout so that the computation is not interrupted.
